+++ releasenotes.md
@@ -3,6 +3,8 @@
 ## Version 0.15.0
 ### New Features
 - Address.CountryCodeLong()
+- Address.Latitude()
+- RandomNumber.NextDouble()
 
 ## Version 0.14.0
 ### New Features